Add Menu, Comparators, Hashset, Rename AvgHeartRate
A Menu system has been added. It utilises a Hashmap with menu names as keys and contents as values.

MainApp has a perform action method which handles the actions that menus perform and carries the user between menus.

Comparators were added in Activity.java using Method Reference.

Hashsets are used in FileScraper.java. This prevents duplicate entries. The hashset is converted to an arraylist to allow for sorting.

avgHeartRate renamed so that variable name matches parameter name.