Add Menus, Filters, Float Input
MainApp starts on GetData menu for user to input CSV filename.

Made the Filter menus interactable.

Added filter interface, with an abstract filter method. The body is defined in MainApp depending on which filter the user chooses.

Filter Sub menus allow the user to set minimum values for distance and duration. These are taken in as a float

TODO finish menus that are dependent on formulas.