Refine Comparable, Add BinarySearch
Comparable now sorts on basis of dates in activity.java

Added method to return an array list of dates to formulae.java

MainApp now loads a default file.

printSearchResults searches activities for a date and tabulates the activity that matches it

Overloaded tabulate activity to take Activity and Activity arraylists.